Feat: add demo search for potential users
