+++ app/components/core/search_component.html.erb
@@ -1,0 +1,79 @@
+<%= content_tag :div, class: 'p-4' do %>
+  <%= form_with url: prepare_itineraries_path do |form| %>
+    <%= content_tag :div, class: 'py-2 text-sm' do %>
+      <%= content_tag :div, class: 'flex flex-wrap justify-between items-center mb-2' do %>
+        <%= search_field_tag  :search, nil,
+                              placeholder: 'Enter a destination',
+                              class: 'w-full sm:w-1/2 xl:w-2/3 text-sm rounded mb-2 sm:m-0',
+                              required: true,
+                              autofocus: true,
+                              autocomplete: 'off',
+                              data: { controller: 'address',
+                                      action: 'google-maps-callback@window->address#initAutocomplete
+                                              keydown->address#keydown' } %>
+
+        <%= form.collection_select :count, count_options, :second, :first, { selected: 3 },
+                                    class: 'w-1/3 sm:w-auto text-sm rounded' %>
+        <span class='m-2 sm:hidden'>results by</span>
+        <%= form.collection_select :sort, sort_options, :second, :first, { selected: 'best' },
+                                    class: 'w-1/3 sm:w-auto text-sm rounded' %>
+      <% end %>
+
+      <%= content_tag :div, class: 'flex items-center', data: { controller: 'date' } do %>
+        <%= date_field_tag  :start, Time.zone.today,
+                            min: Time.zone.today,
+                            class: 'w-1/2 mb-2 text-sm rounded',
+                            autocomplete: 'off',
+                            data: { date_target: 'start', action: 'change->date#updateEndDate' } %>
+
+        <span class='m-2'>to</span>
+        <%= date_field_tag  :end, Time.zone.today,
+                            min: Time.zone.today,
+                            class: 'w-1/2 mb-2 text-sm rounded',
+                            autocomplete: 'off',
+                            data: { date_target: 'end' } %>
+      <% end %>
+
+      <%= content_tag :div, class: 'hidden sm:flex flex-wrap justify-between items-center' do %>
+        <%= content_tag :div, class: 'flex items-center' do %>
+          <span class='hidden md:block font-bold'>BUDGET</span>
+          <%= content_tag :div, class: 'flex text-sm' do %>
+            <% [['$', '1'], ['$$', '1,2'], ['$$$', '1,2,3'], ['$$$$', '1,2,3,4']].each do |(label, value)| %>
+              <%= content_tag :span, class: 'flex items-center' do %>
+                <%= radio_button_tag :budget, value, false, class: 'm-2' %>
+                <%= label_tag label %>
+              <% end %>
+            <% end %>
+          <% end %>
+        <% end %>
+
+        <%= content_tag :div, class: 'flex items-center', data: { controller: 'slider' } do %>
+          <span data-slider-target='output'>15 km</span>
+          <%= range_field_tag :distance, 15, in: 5..40, step: 5, class: 'm-2',
+                              data: { action: 'input->slider#update', slider_target: 'input' } %>
+          <span class='hidden md:block font-bold'>RADIUS</span>
+        <% end %>
+      <% end %>
+    <% end %>
+
+    <%#= render Core::DropdownComponent.new(title: 'Where would you like to stay?') do %>
+      <%#= render Core::CheckboxComponent.new data: lodging_data,
+                                              prefix: 'lodging' %>
+    <%# end %>
+
+    <%= render Core::DropdownComponent.new(title: 'Where would you like to go?') do %>
+      <%= render Core::CheckboxComponent.new data: activity_data,
+                                              prefix: 'activities',
+                                              size: 1 %>
+    <% end %>
+
+    <%= render Core::DropdownComponent.new(title: 'Where would you like to eat?') do %>
+      <%= render Core::CheckboxComponent.new data: restaurant_data,
+                                              prefix: 'restaurants',
+                                              size: 1 %>
+    <% end %>
+
+    <p id='search-btn' class='py-2 text-center'><%= form.submit 'SKEDADDLE', data: { turbo: false },
+                                                                              class: 'blue-btn mb-2' %></p>
+  <% end %>
+<% end %>
