+++ app/controllers/home_controller.rb
@@ -5,6 +5,7 @@ class HomeController < ApplicationController
 
   def home; end
   def about; end
+  def demo; end
   def contact; end
   def received; end
 end
