+++ app/views/home/demo.html.erb
@@ -1,0 +1,1 @@
+<%= render Core::SearchComponent.new %>
